(move) config plugin to plugins folder
