+++ app/start.ts
@@ -4,13 +4,13 @@ import {
   importmap,
   io,
   manifest,
-  pluginConfig,
   pluginIO,
   runWith,
 } from "@radish/core/effects";
 import { Option } from "@radish/core/utils";
 import { resolve } from "@std/path";
 import {
+  pluginConfig,
   pluginManifest,
   pluginRadish,
   pluginStripTypes,
