+++ core/exports/effects.ts
@@ -9,6 +9,6 @@ export {
   transformerFor,
 } from "../src/effects/effects.ts";
 export { io, pluginIO } from "../src/effects/io.ts";
-export { config, denoConfig, pluginConfig } from "../src/effects/config.ts";
+export { config, denoConfig } from "../src/effects/config.ts";
 export { importmap, importmapHandlers } from "../src/effects/impormap.ts";
 export { hot } from "../src/effects/hot-update.ts";
