+++ core/exports/plugins.ts
@@ -1,3 +1,4 @@
 export { pluginStripTypes } from "../src/plugins/strip-types.ts";
+export { pluginConfig } from "../src/plugins/config.ts";
 export { pluginRadish } from "../src/plugins/plugins.ts";
 export { pluginManifest, updateManifest } from "../src/plugins/manifest.ts";
