+++ core/src/effects/config.ts
@@ -1,9 +1,5 @@
-import { assertExists } from "@std/assert";
-import { existsSync } from "@std/fs";
-import * as JSONC from "@std/jsonc";
-import type { Config, Plugin, ResolvedConfig } from "../types.d.ts";
-import { createEffect, createTransformEffect, handlerFor } from "./effects.ts";
-import { io } from "./io.ts";
+import type { Config, ResolvedConfig } from "../types.d.ts";
+import { createEffect, createTransformEffect } from "./effects.ts";
 
 interface ConfigEffect {
   read: () => ResolvedConfig;
@@ -28,21 +24,4 @@ export const denoConfig = {
    * Returns the parsed deno config and throws if it can't find it
    */
   read: createEffect<() => Record<string, any>>("config/read"),
-};
-
-export const pluginConfig: Plugin = {
-  name: "config-plugin",
-  handlers: [
-    handlerFor(denoConfig.read, async () => {
-      const fileName = ["deno.json", "deno.jsonc"]
-        .find((fileName) => existsSync(fileName));
-      assertExists(fileName, "deno config not found");
-
-      const content = await io.readFile(fileName);
-
-      return fileName.endsWith(".json")
-        ? JSON.parse(content)
-        : JSONC.parse(content);
-    }),
-  ],
 };
